+++ lib/src/services/backend_factory.dart
@@ -8,6 +8,7 @@ import 'platform_media_backend.dart';
 import 'platform_shared_url_service.dart';
 import 'queue_store.dart';
 import 'shared_url_service.dart';
+import 'supporter_service.dart';
 
 bool get _isDesktop =>
     !kIsWeb &&
@@ -41,4 +42,12 @@ QueueStore createQueueStore() {
   }
 
   return SharedPreferencesQueueStore();
+}
+
+SupporterService createSupporterService() {
+  if (kIsWeb) {
+    return SupporterService();
+  }
+
+  return SupporterService(store: SharedPreferencesSupporterStore());
 }